Loggedin user management

+++ controllers/contactController.js
@@ -45,17 +45,24 @@ const getContact = asyncHandler(async(req, res) =>{
 // @access Private
 //@route PUT /api/contacts
 
-const updateContact = asyncHandler(async(req, res) =>{
+const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
-  if(!contact) {
+  if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
-    {new : true}
-  )
+    { new: true }
+  );
+
   res.status(200).json(updatedContact);
 });
 
@@ -69,7 +76,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
-  
+
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to update other contacts")
+;   }
+
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
